✨♻️ (Tailwind) - Add Tailwind in project + Reorganize files in separate folders
