+++ src/utils/codeXViewProvider.ts
@@ -1,0 +1,124 @@
+import * as vscode from "vscode";
+import documentations from "../data/documentations";
+
+export class CodeXViewProvider implements vscode.WebviewViewProvider {
+  public static readonly viewType = "codeX-documentations";
+
+  private _view?: vscode.WebviewView;
+
+  constructor(private readonly _extensionUri: vscode.Uri) {}
+
+  public resolveWebviewView(
+    webviewView: vscode.WebviewView,
+    context: vscode.WebviewViewResolveContext,
+    _token: vscode.CancellationToken
+  ) {
+    this._view = webviewView;
+
+    webviewView.webview.options = {
+      enableScripts: true,
+      localResourceRoots: [this._extensionUri],
+    };
+
+    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
+
+    webviewView.webview.postMessage({
+      type: "setDocumentations",
+      documentations,
+    });
+
+    webviewView.onDidChangeVisibility(() => {
+      if (webviewView.visible) {
+        webviewView.webview.postMessage({
+          type: "setDocumentations",
+          documentations,
+        });
+      }
+    });
+  }
+
+  private _getHtmlForWebview(webview: vscode.Webview) {
+    const scriptUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(
+        this._extensionUri,
+        "src",
+        "components",
+        "Sidebar",
+        "main.js"
+      )
+    );
+
+    const styleResetUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(
+        this._extensionUri,
+        "src",
+        "assets",
+        "styles",
+        "reset.css"
+      )
+    );
+
+    const styleVSCodeUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(
+        this._extensionUri,
+        "src",
+        "assets",
+        "styles",
+        "vscode.css"
+      )
+    );
+
+    const styleTailwindUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(
+        this._extensionUri,
+        "src",
+        "assets",
+        "styles",
+        "tailwind.min.css"
+      )
+    );
+
+    const styleMainUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(
+        this._extensionUri,
+        "src",
+        "components",
+        "Sidebar",
+        "main.css"
+      )
+    );
+
+    const nonce = getNonce();
+
+    return `
+      <!DOCTYPE html>
+        <html lang="en">
+        <head>
+          <meta charset="UTF-8">
+          <script src="https://cdn.tailwindcss.com"></script>
+          <meta http-equiv="Content-Security-Policy" content="default-src 'none'; img-src 'self' data: https:; style-src 'self' 'unsafe-inline' ${webview.cspSource}; script-src 'nonce-${nonce}';">
+          <meta name="viewport" content="width=device-width, initial-scale=1.0">
+          <link href="${styleResetUri}" rel="stylesheet">
+          <link href="${styleTailwindUri}" rel="stylesheet">
+          <link href="${styleVSCodeUri}" rel="stylesheet">
+          <link href="${styleMainUri}" rel="stylesheet">
+          <title>Documentation List</title>
+          <script nonce="${nonce}" src="${scriptUri}"></script>
+        </head>
+        <body>
+          <h1 class="text-3xl font-bold">Documentation List</h1>
+          <div id="documentation-list" class="space-y-4"></div>
+        </body>
+      </html>`;
+  }
+}
+
+function getNonce() {
+  let text = "";
+  const possible =
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+  for (let i = 0; i < 32; i++) {
+    text += possible.charAt(Math.floor(Math.random() * possible.length));
+  }
+  return text;
+}
